+++ test/bigtest/interactors/marc-field-protection/index.js
@@ -1,0 +1,1 @@
+export * from './marc-field-protection';