+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Bugs fixed:
 * Fix rendering qualifier sections with old data in match profiles details (UIDATIMP-481)
 * Fix Inconsistent in Holding schema between UI and Backend (UIDATIMP-596)
+* Create MARC field protection settings screen (UIDATIMP-592)
 
 ## [2.1.1](https://github.com/folio-org/ui-data-import/tree/v2.1.1) (2020-07-09)